+++ src/CCA/Components/ICE/impICE.cc
@@ -708,7 +708,7 @@ void ICE::setupRHS(const ProcessorGroup*,
       patch->getFineLevelPatches(finePatches);
     }
 
-    for(int i=0;i<finePatches.size();i++){
+    for(size_t i=0;i<finePatches.size();i++){
       const Patch* finePatch = finePatches[i];
 
       IntVector l, h, fl, fh;
